padronizando sessao de Login
