+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, CssBaseline, makeStyles } from '@material-ui/core';
 import { ptBR } from '@material-ui/core/locale';
-import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
+import { createTheme, ThemeProvider } from '@material-ui/core/styles';
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
 import CadastroMarca from './pages/CadastroMarca';
 import ListagemMarcas from './pages/ListagemMarca';
@@ -9,7 +9,7 @@ import MenuLateral from './components/MenuLateral/MenuLateral.jsx';
 import './App.css';
 import Login from "./pages/Login";
 
-const muiTheme = createMuiTheme(
+const muiTheme = createTheme(
   {
     palette: {
       primary: {
